+++ src/libtcod/mersenne.h
@@ -38,26 +38,26 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-TCODLIB_API TCOD_random_t TCOD_random_get_instance(void);
-TCODLIB_API TCOD_random_t TCOD_random_new(TCOD_random_algo_t algo);
-TCODLIB_API TCOD_random_t TCOD_random_save(TCOD_random_t mersenne);
-TCODLIB_API void TCOD_random_restore(TCOD_random_t mersenne, TCOD_random_t backup);
-TCODLIB_API TCOD_random_t TCOD_random_new_from_seed(TCOD_random_algo_t algo, uint32_t seed);
-TCODLIB_API void TCOD_random_delete(TCOD_random_t mersenne);
+TCODLIB_API TCOD_Random* TCOD_random_get_instance(void);
+TCODLIB_API TCOD_Random* TCOD_random_new(TCOD_random_algo_t algo);
+TCODLIB_API TCOD_Random* TCOD_random_save(TCOD_Random* mersenne);
+TCODLIB_API void TCOD_random_restore(TCOD_Random* mersenne, TCOD_Random* backup);
+TCODLIB_API TCOD_Random* TCOD_random_new_from_seed(TCOD_random_algo_t algo, uint32_t seed);
+TCODLIB_API void TCOD_random_delete(TCOD_Random* mersenne);
 
-TCODLIB_API void TCOD_random_set_distribution(TCOD_random_t mersenne, TCOD_distribution_t distribution);
+TCODLIB_API void TCOD_random_set_distribution(TCOD_Random* mersenne, TCOD_distribution_t distribution);
 
-TCODLIB_API int TCOD_random_get_int(TCOD_random_t mersenne, int min, int max);
-TCODLIB_API float TCOD_random_get_float(TCOD_random_t mersenne, float min, float max);
-TCODLIB_API double TCOD_random_get_double(TCOD_random_t mersenne, double min, double max);
+TCODLIB_API int TCOD_random_get_int(TCOD_Random* mersenne, int min, int max);
+TCODLIB_API float TCOD_random_get_float(TCOD_Random* mersenne, float min, float max);
+TCODLIB_API double TCOD_random_get_double(TCOD_Random* mersenne, double min, double max);
 
-TCODLIB_API int TCOD_random_get_int_mean(TCOD_random_t mersenne, int min, int max, int mean);
-TCODLIB_API float TCOD_random_get_float_mean(TCOD_random_t mersenne, float min, float max, float mean);
-TCODLIB_API double TCOD_random_get_double_mean(TCOD_random_t mersenne, double min, double max, double mean);
+TCODLIB_API int TCOD_random_get_int_mean(TCOD_Random* mersenne, int min, int max, int mean);
+TCODLIB_API float TCOD_random_get_float_mean(TCOD_Random* mersenne, float min, float max, float mean);
+TCODLIB_API double TCOD_random_get_double_mean(TCOD_Random* mersenne, double min, double max, double mean);
 
 TCODLIB_API TCOD_dice_t TCOD_random_dice_new(const char* s);
-TCODLIB_API int TCOD_random_dice_roll(TCOD_random_t mersenne, TCOD_dice_t dice);
-TCODLIB_API int TCOD_random_dice_roll_s(TCOD_random_t mersenne, const char* s);
+TCODLIB_API int TCOD_random_dice_roll(TCOD_Random* mersenne, TCOD_dice_t dice);
+TCODLIB_API int TCOD_random_dice_roll_s(TCOD_Random* mersenne, const char* s);
 #ifdef __cplusplus
 }
 #endif